# docs/api/openapi-baseline.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RustCI API",
    "version": "1.0.0",
    "description": "A high-performance CI/CD platform built in Rust",
    "contact": {
      "name": "RustCI Team",
      "email": "support@rustci.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.rustci.dev",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/cluster/nodes": {
      "get": {
        "tags": ["cluster"],
        "summary": "List cluster nodes",
        "description": "Returns a list of all cluster nodes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of cluster nodes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNodesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "tags": ["cluster"],
        "summary": "Join node to cluster",
        "description": "Adds a new node to the cluster",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinNodeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Node joined successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinNodeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/cluster/status": {
      "get": {
        "tags": ["cluster"],
        "summary": "Get cluster status",
        "description": "Returns the current status of the cluster",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cluster status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for API authentication. Format: Bearer <token>"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["error", "status", "timestamp"]
      },
...
      "JoinNodeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["Master", "Worker"]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": ["name", "address", "role"]
      },
      "JoinNodeResponse": {
        "type": "object",
        "properties": {
          "node": {
            "$ref": "#/components/schemas/ClusterNode"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ClusterNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["Master", "Worker"]
          },
          "status": {
            "type": "string",
            "enum": ["Active", "Inactive", "Failed"]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_heartbeat": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ListNodesResponse": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClusterNode"
            }
          },
          "total": {
            "type": "integer"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
      "ClusterStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["Healthy", "Degraded", "Unhealthy"]
          },
          "metrics": {
            "$ref": "#/components/schemas/ClusterMetrics"
          },
          "info": {
            "$ref": "#/components/schemas/ClusterInfo"
          }
        }
      },
      "ClusterMetrics": {
        "type": "object",
        "properties": {
          "total_nodes": {
            "type": "integer"
          },
          "healthy_nodes": {
            "type": "integer"
          },
          "failed_nodes": {
            "type": "integer"
          }
        }
      },
      "ClusterInfo": {
        "type": "object",
        "properties": {
          "cluster_id": {
            "type": "string"
          },
          "cluster_name": {
            "type": "string"
          },
          "total_nodes": {
            "type": "integer"
          },
          "active_nodes": {
            "type": "integer"
          },
          "failed_nodes": {
            "type": "integer"
          },
          "uptime_seconds": {
            "type": "integer"
          }
        }
      }
    }
  }
}